Refactor components into directories
