+++ acs-admin/src/pages/AccessControl/AccessControl.vue
@@ -49,8 +49,8 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
 import { defineAsyncComponent } from 'vue'
 import { usePrincipalStore } from '@store/usePrincipalStore.js'
 import { useGroupStore } from '@store/useGroupStore.js'
-import PrincipalList from './PrincipalList.vue'
-import GroupList from './GroupList.vue'
+import PrincipalList from './Principals/PrincipalList.vue'
+import GroupList from './Groups/GroupList.vue'
 import PermissionList from '@pages/AccessControl/Permissions/PermissionList.vue'
 import { usePermissionStore } from '@store/usePermissionStore.js'
 import { UUIDs } from '@amrc-factoryplus/service-client'
